Get rid of is_python_min
this builds off https://github.com/conda-forge/conda-forge-pinning-feedstock/pull/7474

To get rid of this zip key.

It is actually error prone when users overwrite the python_min entry
with jinja syntax so I feel like we shouldn't be recommending something
that will keep things inconsistent.

PRs made to feedstocks that use this:

- [ ] NEED TO FIND THE FEW RECIPES FOR THIS